Eric Sunshine made the following changes:
-*- Renamed celData::SetBool() to Set(), and celDataBufHelper::SetBool()
    to Set() in order to fix numerous ambiguous invocations of Set(bool)
    throughout the project.  The SetBool() variations are no longer needed
    since we no longer support old compilers lacking a bool type.

-*- Eliminated numerous warnings about incorrect printf() formatting
    directive mismatches.

+++ plugins/stdphyslayer/pl.cpp
@@ -185,8 +185,8 @@ void celPlLayer::RemoveEntity (iCelEntity *entity)
   {
     csReport (object_reg, CS_REPORTER_SEVERITY_ERROR,
 	"crystalspace.cel.pllayer",
-	"error while removing Entity with ID %u (%s)", entity->GetID(),
-	entity->GetName());
+	"error while removing Entity with ID %u (%s)",
+	(unsigned int)entity->GetID(), entity->GetName());
     return;
   }
